refactor: 전역에서 사용하는 toast 컴포넌트 상태를 store로 관리
