+++ src/hooks/useToast.ts
@@ -1,4 +1,4 @@
-import { useToastStore } from '@/store/toast';
+import { useToastStore } from '@/stores/toast';
 import { v4 } from 'uuid';
 import type { ToastItemType } from '@/components/ui/Toast';
